refactor(language): relocate lint framework infra into core/analysis
Four moves consolidating lint infrastructure where it belongs:

  core/analysis/lint.ts              → core/analysis/lint-engine.ts
  lint/schema-walker.ts              → core/analysis/schema-walker.ts
  lint/symbol-table.ts               → core/analysis/symbol-table.ts
  lint/position-index.ts             → core/analysis/position-index-pass.ts

The file formerly called `lint.ts` contains the `LintEngine`/`PassStore`/
`defineRule` framework — not lint rules. Rename makes that explicit.

`schema-walker.ts`, `symbol-table.ts`, and `position-index.ts` under `lint/`
were framework adapters (e.g., the `position-index.ts` pass populates the
key already defined in `core/analysis/position-index.ts`), not lint rules.
Moving them next to their key/interface definitions removes an awkward
split where `service.ts` imported the key from `core/analysis/` while the
pass that populated it lived in `lint/`. The `lint/` directory is now
purely rules + utilities.

Also make `core/analysis/index.ts` the canonical barrel for analysis
exports — it now covers what the package-level `src/index.ts` was
re-exporting via deep paths. Renamed position-index-pass.ts to avoid
clashing with the existing position-index.ts that holds the `storeKey` and
query helpers.

All imports across the package updated; full typecheck + test suite green.
