+++ packages/language/src/lint/duplicate-keys.test.ts
@@ -10,7 +10,7 @@ import { parse } from '@agentscript/parser';
 import { Dialect } from '../core/dialect.js';
 import { Block, NamedBlock, NamedCollectionBlock } from '../core/block.js';
 import { StringValue } from '../core/primitives.js';
-import { LintEngine } from '../core/analysis/lint.js';
+import { LintEngine } from '../core/analysis/lint-engine.js';
 import { createSchemaContext } from '../core/analysis/scope.js';
 import { duplicateKeyPass } from './duplicate-keys.js';
