+++ packages/language/src/lint/duplicate-keys.ts
@@ -12,7 +12,7 @@ import {
   storeKey,
   type LintPass,
   type PassStore,
-} from '../core/analysis/lint.js';
+} from '../core/analysis/lint-engine.js';
 import { FieldChild, StatementChild } from '../core/children.js';
 import { isSetClause, isWithClause } from '../core/guards.js';
 import { lintDiagnostic } from './lint-utils.js';
