+++ packages/language/src/lint/singular-collection.ts
@@ -13,7 +13,7 @@ import {
   schemaContextKey,
   type LintPass,
   type PassStore,
-} from '../core/analysis/lint.js';
+} from '../core/analysis/lint-engine.js';
 import { lintDiagnostic } from './lint-utils.js';
 
 /**
