+++ packages/language/src/lint/unreachable-code.ts
@@ -14,7 +14,7 @@ import {
   storeKey,
   type LintPass,
   type PassStore,
-} from '../core/analysis/lint.js';
+} from '../core/analysis/lint-engine.js';
 import { lintDiagnostic } from './lint-utils.js';
 import {
   type Statement,
