+++ packages/language/src/parse-and-lint.ts
@@ -13,7 +13,7 @@
 import type { AstRoot } from './core/types.js';
 import type { Diagnostic } from './core/diagnostics.js';
 import { Dialect } from './core/dialect.js';
-import { LintEngine, PassStore } from './core/analysis/lint.js';
+import { LintEngine, PassStore } from './core/analysis/lint-engine.js';
 import { createSchemaContext } from './core/analysis/scope.js';
 import type { DialectConfig } from './dialect-config.js';
 import type { SyntaxNode } from './core/types.js';
